refactor(api): reorganize namespaces and improve exception handling in HeadlessApiExceptionHandler
